# docs/specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LoonFS HTTP API",
    "description": "Static OpenAPI document for the LoonFS v0 HTTP API.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.2.1"
  },
  "paths": {
...
    "/readiness": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Check readiness",
        "description": "Returns `ready` while the server admits new work. Once shutdown begins and publisher admission closes, answers 503 `shutting_down` so load balancers can drain the instance. `/health` stays the liveness probe: it only reports that the process is up.",
        "operationId": "readiness",
        "responses": {
          "200": {
            "description": "The server admits new work",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "503": {
            "description": "Shutdown has begun; admission is closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {}
        ]
      }
    },
    "/v0/admin/namespaces/{namespace_id}/checkpoints": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List checkpoints",
        "description": "Lists one page of active checkpoints in checkpoint-id order. Expired checkpoints remain visible until collection releases them. Released checkpoints are omitted. The cursor resumes a live listing and does not create a snapshot.",
        "operationId": "list_checkpoints",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1000,
              "maximum": 1000,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque checkpoint-list page cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active checkpoint objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCheckpointsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid namespace id, limit, or cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create checkpoint",
        "description": "Creates a named, user-owned checkpoint record pinning the current namespace view. Every call mints a new record under a new id; the name is a label, not a key. The record is a garbage-collection root until it is released, so routine maintenance should flush the WAL instead. This is a maintenance/admin operation, not a file mutation.",
        "operationId": "create_checkpoint",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Checkpoint name and optional lifetime",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Namespace envelope containing the created checkpoint",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCheckpointResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid namespace id, name, or lifetime",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "410": {
            "description": "Namespace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Checkpoint unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/admin/namespaces/{namespace_id}/checkpoints/{checkpoint_id}/release": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Release checkpoint",
        "description": "Releases a user-owned checkpoint pin by id. Idempotent: releasing an already-released or reaped record succeeds. The record is reaped by a later garbage-collection pass; its pinned data becomes collectable only on the pass after that.",
        "operationId": "release_checkpoint",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkpoint_id",
            "in": "path",
            "description": "Checkpoint id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Checkpoint release accepted (including an already released or reaped checkpoint)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseCheckpointResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id, or the checkpoint is fork-owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          }
        }
      }
    },
    "/v0/admin/namespaces/{namespace_id}/diagnostics": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get namespace diagnostics",
        "description": "Returns namespace state together with the current manifest and visible WAL tail.",
        "operationId": "get_namespace_diagnostics",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Namespace diagnostics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceDiagnostics"
                }
              }
            }
          },
          "400": {
            "description": "Invalid namespace id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "410": {
            "description": "Namespace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/admin/namespaces/{namespace_id}/grep/index": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Read the grep index's lifecycle",
        "description": "Reports where the namespace's grep index is: `disabled`, `backfilling` with the sequence it walks toward and how far the walk got, or `active` with the watermark it has built through. One grep root read, no side effects. A namespace that never enabled the index reads as `disabled`. Requires this deployment to maintain the grep index.",
        "operationId": "get_grep_index_status",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The index's lifecycle and bookkeeping",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrepIndexStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The backing store rejected its configured credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "500": {
            "description": "The grep index is corrupt or its backing store is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "This deployment does not maintain the grep index",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v0/capabilities": {
      "get": {
        "tags": [
          "capabilities"
        ],
        "summary": "Get capabilities",
        "description": "Returns a summary of supported features and limits.",
        "operationId": "capabilities",
        "responses": {
          "200": {
            "description": "Capability document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityDocument"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "HTTP error body used by LoonFS APIs.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable reason from the [`ErrorCode`](crate::ErrorCode)\nregistry.\n\nCarried as a string so clients keep working when a newer server\nintroduces a code they do not know; use\n[`ErrorCode::parse`](crate::ErrorCode::parse) for typed access."
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorDetails",
                "description": "Structured context for the code, present when the failure carries\nmachine-usable identity (API spec, \"Standard error contract\"). Boxed\nso the rare detailed error does not widen every error-carrying result."
              }
            ]
          },
          "feature": {
            "type": [
              "string",
              "null"
            ],
            "description": "For `not_supported` errors, the capability-document feature key the\nclient should reconcile against."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          },
          "param": {
            "type": [
              "string",
              "null"
            ],
            "description": "Identifies the invalid input. Body fields use JSON Pointer paths;\nquery and path parameters use their names; CLI errors use the flag or\nargument as written."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation id the server assigned to the failed request; the same\nvalue is sent as the `x-request-id` response header."
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The deployment's `auth_token`, sent as `Authorization: Bearer <token>`. A server configured without a token accepts every request; one configured with a token answers 401 `unauthorized` without it."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}